backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
requests==2.32.3
python-dotenv==1.0.1
gunicorn==22.0.0
openai==1.38.0
python-docx==1.1.2
pdfplumber==0.11.2